refactor: Remove dynamic definition of DatabagModel
Using a dynamic definition makes it impossible for a static type checker to validate it. This avoids some new `pyright` errors and allows us to update the `uv.lock`.

The disadvantage to this approach is that we stray a bit from the definitions of the classes that are (very unfortunately) copied around between charms.

An alternative would be to ignore the warnings, but that's usually a last resort.